Completed statistical generators in DimensionalSpace
Changed number of points variable in DimensionalSpace (previously n) to an integer, and renamed it to numberOfPoints for clarity.

Finished the functions to calculate the sums, means, and standard deviations for the points. Also added in a function  to oversee those functions. Added a call to this function at the end of addPts to automate.