Reorganized Files
Created a new Archives folder within the File folder, to hold all files related to past milestones. Renamed many files for consistency's sake.

From now on, files related to the milestone (such as user manual, UML, etc.) are place in the File folder. When the milestone is completed, move these files and all classes to a new Archive file